fix: コンポーネントの命名ミス
